Use double precision when converting duration.
Otherwise there will be truncation after a couple of days.

+++ src/rules/simulation/WorldRouter.cpp
@@ -262,7 +262,7 @@ void WorldRouter::deliverTo(const Operation& op, Ref<LocatedEntity> ent)
     }
     //Set the time of when this op is dispatched. That way, other components in the system can
     //always use the seconds set on the op to know the current time.
-    op->setSeconds(std::chrono::duration_cast<std::chrono::duration<float>>(getTime()).count());
+    op->setSeconds(std::chrono::duration_cast<std::chrono::duration<double>>(getTime()).count());
 
     OpVector res;
     debug(std::cout << "WorldRouter::deliverTo begin {"
